changed around dashboard card
