+++ src/components/dashboard/event-view/recent-card/header/Header.jsx
@@ -1,17 +1,55 @@
-import React from 'react';
-import { Link, useHistory } from 'react-router-dom';
+import React, { useState } from 'react';
+import { Link } from 'react-router-dom';
 import { cardStyles } from '../../../../../styles';
 import CardHeader from '@material-ui/core/CardHeader';
 import Avatar from '@material-ui/core/Avatar';
 import Typography from '@material-ui/core/Typography';
+import { print } from 'graphql';
 import { Icon } from '@iconify/react';
+import starEmpty from '@iconify-icons/dashicons/star-empty';
+import starFilled from '@iconify-icons/carbon/star-filled';
 import { timeAgo } from '../../../../../utilities/functions';
-import softwareUpload from '@iconify-icons/gg/software-upload';
+import { axiosWithAuth } from '../../../../../utilities/axiosWithAuth';
+import { ADD_FAVORITE_EVENT, REMOVE_FAVORITE_EVENT } from '../../../../../graphql/users/user-mutations';
 
 const Header = (props) => {
+  const [favorite, setFavorite] = useState(props.isFavorite);
   const classes = cardStyles();
   const shownTime = timeAgo(props.createDateTime);
-  const { push } = useHistory();
+
+  const addFavoriteEvent = () => {
+    const favoriteEvent = {
+      event_id: Number(props.id),
+      user_id: Number(props.currentUserId),
+    };
+
+    axiosWithAuth()
+      .post(`${process.env.REACT_APP_BASE_URL}/graphql`, {
+        query: print(ADD_FAVORITE_EVENT),
+        variables: { favoriteEvent: favoriteEvent },
+      })
+      .then(() => {
+        setFavorite(!favorite);
+      })
+      .catch((err) => console.dir(err));
+  };
+
+  const removeFavoriteEvent = () => {
+    const favoriteEvent = {
+      event_id: Number(props.id),
+      user_id: Number(props.currentUserId),
+    };
+
+    axiosWithAuth()
+      .post(`${process.env.REACT_APP_BASE_URL}/graphql`, {
+        query: print(REMOVE_FAVORITE_EVENT),
+        variables: { favoriteEvent: favoriteEvent },
+      })
+      .then((res) => {
+        setFavorite(!favorite);
+      })
+      .catch((err) => console.dir(err));
+  };
 
   return (
     <>
@@ -23,6 +61,7 @@ const Header = (props) => {
               title={`${props.User.firstName} ${props.User.lastName}`}
               aria-label="avatar"
               className={classes.avatar}
+              style={{ height: '55px', width: '55px' }}
               src={props.User.photo ? props.User.photo : ''}
             >
               {!props.User.photo && (
@@ -35,7 +74,7 @@ const Header = (props) => {
         }
         title={
           <div className={classes.titleContainer}>
-            <div>
+            <div style={{ display: 'flex', width: '85%', alignItems: 'center' }}>
               <Link to={`/profile/${props.User.id}`} style={{ cursor: 'pointer' }}>
                 <Typography variant="h6">{`${props.User.firstName} `}</Typography>
               </Link>
@@ -43,9 +82,15 @@ const Header = (props) => {
                 &nbsp;created an event
               </Typography>
             </div>
-            <button className={classes.dashboardNavigateBtn} onClick={() => push(`events/${props.id}`)}>
-              <Icon icon={softwareUpload} />
-            </button>
+            {!favorite ? (
+              <div className={classes.headerBtn} onClick={addFavoriteEvent}>
+                <Icon icon={starEmpty} />
+              </div>
+            ) : (
+              <div className={classes.headerBtn} onClick={removeFavoriteEvent}>
+                <Icon icon={starFilled} />
+              </div>
+            )}
           </div>
         }
         subheader={
